cricketPro/requirements.txt
astroid==2.3.3
Django==2.2.7
django-bootstrap3==11.1.0
django-ckeditor==5.8.0
django-crispy-forms==1.8.1
django-js-asset==1.2.2
django-nested-inline==0.3.7
djangorestframework==3.10.3
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==6.2.1
pkg-resources==0.0.0
pylint==2.4.3
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
typed-ast==1.4.0
wrapt==1.11.2
